# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/ambulances/{id}/statuses": {
      "post": {
        "operationId": "createAmbulanceStatus",
        "summary": "Create a status for an ambulance",
        "tags": [
          "AmbulanceStatus"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "QAP",
                      "OCP",
                      "EVT",
                      "J4",
                      "J5",
                      "FA",
                      "MN"
                    ]
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/bases/{baseId}/phones": {
      "post": {
        "operationId": "createBasePhoneNumber",
        "summary": "Create base phone number",
        "tags": [
          "Base"
        ],
        "description": "Create a phone number for a base",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "description": "Brazilian phone number (example: +5511999999999)",
                    "type": "string"
                  }
                },
                "required": [
                  "number"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "baseId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/companies/{companyId}/phones": {
      "post": {
        "operationId": "createCompanyPhoneNumber",
        "summary": "Create company phone number",
        "tags": [
          "Company"
        ],
        "description": "Create company phone number",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "description": "Brazilian phone number (example: +5511999999999)",
                    "type": "string"
                  }
                },
                "required": [
                  "number"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/{chatId}": {
      "get": {
        "operationId": "findMessagesByChat",
        "summary": "List messages by chat with pagination",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "perPage",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "chatId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "chatId": {
                            "type": "string"
                          },
                          "messageContent": {
                            "type": "string"
                          },
                          "messageType": {
                            "type": "string"
                          },
                          "messageFile": {
                            "nullable": true,
                            "type": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "avatarUrl": {
                                "nullable": true,
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "avatarUrl"
                            ],
                            "additionalProperties": false
                          },
                          "messageReadReceipt": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "readAt": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              },
                              "required": [
                                "userId",
                                "readAt"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "chatId",
                          "messageContent",
                          "messageType",
                          "messageFile",
                          "user",
                          "messageReadReceipt",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPreviousPage": {
                          "type": "boolean"
                        },
                        "currentPage": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "total",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage",
                        "currentPage"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/chats/{chatId}/read": {
      "patch": {
        "operationId": "readMessages",
        "summary": "Mark chat messages as read",
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "chatId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markedCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "markedCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}